Rework tracks grouping/presentation
As per our last syllabus team meeting, we think these are the three
families of tracks we have.

Explain a bit of context as to what each category is. This maybe makes
more sense to do under the headings of each group of tracks? But then
the text is going to be awfully skinny...

Also, self-study tracks may have specific prereqs - we should list them
so people know what they may want to opt to study.